fix: schemas

+++ src/schemas/movies.schema.ts
@@ -5,7 +5,7 @@ const movieSchemas = z.object({
   name: z.string().max(50),
   description: z.string().nullish(),
   duration: z.number().positive(),
-  price: z.number().positive(),
+  price: z.number().positive().int(),
 });
 
 const movieCreateSchemasOmit = movieSchemas.omit({ id: true });
